pyproject.toml
[tool.poetry]
name = "pytexresume"
version = "0.2.0"
description = "A Python-based resume generator with LaTeX support."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "pytexresume" }]


[tool.poetry.scripts]
pytexresume = "pytexresume.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
PyYaml = "*"
gitpython = "*"
pylatex = "*"
pydantic = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
